chore(market-matcher): refactor to separate matcher
